New feature: GekkoFS Proxy

The GekkoFS proxy is an additional (alternative) component that runs on each client and acts as gateway between the
client and daemons. It can improve network stability, e.g., for opa-psm2, and provides a basis for future asynchronous
I/O as well as client caching techniques to control file system semantics.

The `gkfs` script fully supports the GekkoFS proxy and an example can be found in `scripts/run`. When using the proxy
manually additional arguments are required on the daemon side, i.e., which network interface and protocol should be
used:

```bash
<daemon args> --proxy-listen eno1 --proxy-protocol ofi+sockets
```

The proxy is started thereafter:

```bash
./gkfs_proxy -H ./gkfs_hostfile --pid-path ./vef_gkfs_proxy.pid -p ofi+sockets
```

The shared hostfile was generated by the daemons whereas the pid_path is local to the machine and is
detected by clients. The pid-path defaults to `/tmp/gkfs_proxy.pid`.

Under default operation, clients detect automatically whether to use the proxy. If another proxy path is used, the
environment variable `LIBGKFS_PROXY_PID_FILE` can be set for the clients.

Alternatively, the `gkfs` automatically sets all required arguments:

```bash
scripts/run/gkfs -c scripts/run/gkfs.conf -f start --proxy
* [gkfs] Starting GekkoFS daemons (1 nodes) ...
* [gkfs] GekkoFS daemons running
* [gkfs] Startup time: 2.013 seconds
* [gkfs] Starting GekkoFS proxies (1 nodes) ...
* [gkfs] GekkoFS proxies running
* [gkfs] Startup time: 5.002 seconds
Press 'q' to exit
```

Please consult `include/config.hpp` for additional configuration options. Note, GekkoFS proxy does not support
replication.

Closes https://storage.bsc.es/gitlab/hpc/gekkofs/-/issues/114

Closes #114

See merge request hpc/gekkofs!191

+++ src/daemon/handler/srv_metadata.cpp
@@ -253,6 +253,11 @@ rpc_srv_remove_metadata(hg_handle_t handle) {
                 GKFS_DATA->storage()->destroy_chunk_space(in.path);
         }
 
+    } catch(const gkfs::metadata::NotFoundException& e) {
+        GKFS_DATA->spdlogger()->warn(
+                "{}(): path '{}' message '{}'. Continuing, setting out.err 0.",
+                __func__, in.path, e.what());
+        out.err = 0;
     } catch(const gkfs::metadata::DBException& e) {
         GKFS_DATA->spdlogger()->error("{}(): path '{}' message '{}'", __func__,
                                       in.path, e.what());
@@ -311,7 +316,8 @@ rpc_srv_remove_data(hg_handle_t handle) {
 
     // Remove all chunks for that file
     try {
-        GKFS_DATA->storage()->destroy_chunk_space(in.path);
+        if(!gkfs::config::limbo_mode)
+            GKFS_DATA->storage()->destroy_chunk_space(in.path);
         out.err = 0;
     } catch(const gkfs::data::ChunkStorageException& e) {
         GKFS_DATA->spdlogger()->error(
@@ -435,8 +441,9 @@ rpc_srv_update_metadentry_size(hg_handle_t handle) {
         out.err = EBUSY;
     }
 
-    GKFS_DATA->spdlogger()->debug("{}() Sending output '{}'", __func__,
-                                  out.err);
+    GKFS_DATA->spdlogger()->debug(
+            "{}() Sending output err '{}' ret_offset '{}'", __func__, out.err,
+            out.ret_offset);
     auto hret = margo_respond(handle, &out);
     if(hret != HG_SUCCESS) {
         GKFS_DATA->spdlogger()->error("{}() Failed to respond", __func__);
@@ -487,8 +494,8 @@ rpc_srv_get_metadentry_size(hg_handle_t handle) {
         out.err = EBUSY;
     }
 
-    GKFS_DATA->spdlogger()->debug("{}() Sending output '{}'", __func__,
-                                  out.err);
+    GKFS_DATA->spdlogger()->debug("{}() Sending output err '{}' ret_size '{}'",
+                                  __func__, out.err, out.ret_size);
     auto hret = margo_respond(handle, &out);
     if(hret != HG_SUCCESS) {
         GKFS_DATA->spdlogger()->error("{}() Failed to respond", __func__);
